python/genius_autogen/requirements.txt
pyautogen==0.1.14
